api/sample_request.py
import googleapiclient.discovery
import sys

from face_detector.interface import FaceDetectorInterface


def predict_json(project, model, instances, version=None):
    """Send json data to a deployed model for prediction.

    Args:
        project (str): project where the Cloud ML Engine Model is deployed.
        model (str): model name.
        instances ([Mapping[str: Any]]): Keys should be the names of Tensors
            your deployed model expects as inputs. Values should be datatypes
            convertible to Tensors, or (potentially nested) lists of datatypes
            convertible to tensors.
        version: str, version of the model to target.
    Returns:
        Mapping[str: any]: dictionary of prediction results defined by the
            model.
    """
    # Create the ML Engine service object.
    # To authenticate set the environment variable
    # GOOGLE_APPLICATION_CREDENTIALS=<path_to_service_account_file>
    service = googleapiclient.discovery.build('ml', 'v1')
    name = 'projects/{}/models/{}'.format(project, model)

    if version is not None:
        name += '/versions/{}'.format(version)

    response = service.projects().predict(
        name=name,
        body={'instances': instances}
    ).execute(num_retries=2)

    if 'error' in response:
        raise RuntimeError(response['error'])

    return response['predictions']


if __name__ == "__main__":
    img_path = '../face_detector/input_images/small_img.png'
    img = FaceDetectorInterface.read_image(img_path)
    img = img[:200, :200, :]  # Max bytes on instances is a huge burden

    project = 'stoked-brand-285120'
    model = 'face_detector'
    instances = [img.tolist()]

    res = predict_json(project, model, instances)
    print('Output:', res)
