requirements.txt
opencv-python>=4.4.0.46
torch>=1.6.0
numpy>=1.19.1
torchvision>=0.7.0
matplotlib>=3.3.1
Pillow>=8.0.1